Refactor: Moved parsing of Xing and Info tags together
Since Xing and Info tags are basically the same, there is no need to
differentiate between them when parsing.

The only difference is that "Info" tags should be used for CBR
and "Xing" tags should be used for VBR.

Improved error message when magic bytes cannot be found when looking for
"Xing" or "Info" tags.